Move HoodieAvroReader to hoodie-common, it will be used for compaction and in the record reader
